bug/kernel/string: Fixed string bug
Fixed a bug in the strncpy function. The strncpy function was null terminated which was wrong.

Create a new function 'strcpym' that is null terminated. So it copys at most a m (max) amount of chars.

Replaced the strncpy with the strcpym function in places where we expected a null returned string from the function. This change also fixed the filesystem again