feature/kernel/fatCompability: (#85) (#79)
We now support MBR partitioning of disks and seperate a physical disk into Volumes (a partition).
The Volume indexing has changed to a VOLUMELABEL indexing. A VOLUMELABEL is a character. By thefault the filesystem of the kernel is always given the 'C' drive letter.
Paths now need to work with a windows like C:/... indexing.

The default driveletter can be chaged with the BOBAOS_DEFAULT_VOLUME_LABEL entry in the config.
The max paritions per disks is by default 4 (because we only support MBR). This can also be changed in the config with the BOBAOS_MAX_PARTITIONS_PER_DISK entry.

A new partitionTracer was created that is build ontop of the diskStream. With this a filesystem can navigate a parition without the need to calculate the relative addresses of the filesystem to the absolute addresses of the disk.