feature/kernel/fatCompatibility: (#86) (#83) (#80)
The bug (#80) is fixed.

The pathTracer was updated to work on partitions and with the new driveLetter system (#86).
Also a bug in the 'pathTracerGetPathString' function was fixed.

We now support partitioning. So we moved the FAT32 parition of the OS-Disk up to 1MB and now we can mount the fileSystem on MacOS and Windows. (#83).

But because the OS is case-sensitive but FAT is it not per specification, macOS and windows may not create a LFN entry for names shorter than 8 characters.

In macOS if we name a file with a uppercase letter at the start, macOS creates a LFN entry.

And we still need to add support for 'chained' LFN entries lol.